LayoutTests/accessibility/mac/search-with-frames-expected.txt


This tests whether searching works to find frames and to find elements within frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resultElement.role is 'AXRole: AXButton'
PASS resultElement.title is 'AXTitle: button1'
PASS resultElement.role is 'AXRole: AXButton'
PASS resultElement.title is 'AXTitle: button2'
PASS resultElement.role is 'AXRole: AXButton'
PASS resultElement.title is 'AXTitle: button3'
PASS resultElement.role is 'AXRole: AXButton'
PASS resultElement.title is 'AXTitle: button2'
PASS resultElement.role is 'AXRole: AXWebArea'
PASS resultElement.description is 'AXDescription: frame1'
PASS resultElement.role is 'AXRole: AXWebArea'
PASS resultElement.description is 'AXDescription: frame1a'
PASS resultElement.role is 'AXRole: AXWebArea'
PASS resultElement.description is 'AXDescription: frame2'
PASS successfullyParsed is true

TEST COMPLETE
